Use pretty from upstream package
Gustavo's fork (github.com/niemeyer/pretty) contains additional fixes on
recursive elements, so that they don't hang the printer. This is tracked
as issue https://github.com/kr/pretty/issues/72

While the extra patches in the fork are nice, they are not essential and
not needing the fork makes packaging easier.

Signed-off-by: Zygmunt Krynicki <zygmunt.krynicki@huawei.com>